requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
djangorestframework==3.14.0
idna==3.4
Pillow==10.0.0
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
svgwrite==1.4.3
typing-extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4